Create pm_plot_maps.R
This script plots maps showing the spatial distribution of resident and wide-ranging behaviour for each of the five study species and two  life-stages. The only requirement is to set the working directory, which should include a sub-folder called 'data', with a data set called 'map_data.csv'. All plots are exported to 'output' sub-folder.